Additions/x11/vboxvideo: fix the unwanted sign extension of the framebuffer address on 64bit XFree86 systems
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@25486 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/x11/vboxvideo/vboxvideo_68.h
@@ -134,7 +134,7 @@ typedef struct _VBOXRec
     pciVideoPtr pciInfo;
     PCITAG pciTag;
     CARD16 maxBytesPerScanline;
-    int mapPhys, mapOff, mapSize;	/* video memory */
+    unsigned long mapPhys, mapOff, mapSize;	/* video memory */
     void *base, *VGAbase;
     CARD8 *state, *pstate;	/* SVGA state */
     int statePage, stateSize, stateMode;
